+++ oolong/src/commonMain/kotlin/oolong/runtime.kt
@@ -13,11 +13,24 @@ import kotlin.jvm.JvmOverloads
  * Create a runtime.
  */
 @JvmOverloads
+fun <Model, Msg> runtime(
+    init: () -> Pair<Model, Effect<Msg>>,
+    update: (Msg, Model) -> Pair<Model, Effect<Msg>>,
+    view: (Model, Dispatch<Msg>) -> Any?,
+    runtimeContext: CoroutineContext = Dispatchers.Default,
+    renderContext: CoroutineContext = Dispatchers.Main,
+    effectContext: CoroutineContext = Dispatchers.Default
+): Job = runtime(init, update, { it }, view, runtimeContext, renderContext, effectContext)
+
+/**
+ * Create a runtime.
+ */
+@JvmOverloads
 fun <Model, Msg, Props> runtime(
-    init: Init<Model, Msg>,
-    update: Update<Model, Msg>,
-    view: View<Model, Props>,
-    render: Render<Msg, Props>,
+    init: () -> Pair<Model, Effect<Msg>>,
+    update: (Msg, Model) -> Pair<Model, Effect<Msg>>,
+    view: (Model) -> Props,
+    render: (Props, Dispatch<Msg>) -> Any?,
     runtimeContext: CoroutineContext = Dispatchers.Default,
     renderContext: CoroutineContext = Dispatchers.Main,
     effectContext: CoroutineContext = Dispatchers.Default
@@ -45,10 +58,10 @@ object Oolong {
         )
     )
     fun <Model, Msg, Props> runtime(
-        init: Init<Model, Msg>,
-        update: Update<Model, Msg>,
-        view: View<Model, Props>,
-        render: Render<Msg, Props>,
+        init: () -> Pair<Model, Effect<Msg>>,
+        update: (Msg, Model) -> Pair<Model, Effect<Msg>>,
+        view: (Model) -> Props,
+        render: (Props, Dispatch<Msg>) -> Any?,
         runtimeContext: CoroutineContext = Dispatchers.Default,
         renderContext: CoroutineContext = Dispatchers.Main,
         effectContext: CoroutineContext = Dispatchers.Default
@@ -63,10 +76,10 @@ object Oolong {
 }
 
 private class RuntimeImpl<Model, Msg, Props>(
-    init: Init<Model, Msg>,
-    private val update: Update<Model, Msg>,
-    private val view: View<Model, Props>,
-    private val render: Render<Msg, Props>,
+    init: () -> Pair<Model, Effect<Msg>>,
+    private val update: (Msg, Model) -> Pair<Model, Effect<Msg>>,
+    private val view: (Model) -> Props,
+    private val render: (Props, Dispatch<Msg>) -> Any?,
     private val runtimeContext: CoroutineContext,
     private val renderContext: CoroutineContext,
     private val effectContext: CoroutineContext
@@ -91,7 +104,7 @@ private class RuntimeImpl<Model, Msg, Props>(
         }
     }
 
-    private fun step(next: Next<Model, Msg>) {
+    private fun step(next: Pair<Model, Effect<Msg>>) {
         val (state, effect) = next
         val props = view(state)
         currentState = state